+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/ASTMethodReference.java
@@ -10,6 +10,7 @@ import org.checkerframework.checker.nullness.qual.Nullable;
 import net.sourceforge.pmd.lang.java.symbols.JConstructorSymbol;
 import net.sourceforge.pmd.lang.java.types.JMethodSig;
 import net.sourceforge.pmd.lang.java.types.JTypeMirror;
+import net.sourceforge.pmd.lang.java.types.OverloadSelectionResult;
 import net.sourceforge.pmd.lang.java.types.TypeSystem;
 
 /**
@@ -29,7 +30,7 @@ public final class ASTMethodReference extends AbstractJavaExpr
                FunctionalExpression {
 
     private JMethodSig functionalMethod;
-    private JMethodSig compileTimeDecl;
+    private OverloadSelectionResult compileTimeDecl;
 
     private String methodName;
 
@@ -152,6 +153,11 @@ public final class ASTMethodReference extends AbstractJavaExpr
      * @see #getTypeMirror()
      */
     public JMethodSig getReferencedMethod() {
+        return getOverloadSelectionInfo().getMethodType();
+    }
+
+    @Override
+    public OverloadSelectionResult getOverloadSelectionInfo() {
         forceTypeResolution();
         return assertNonNullAfterTypeRes(compileTimeDecl);
     }
@@ -160,8 +166,8 @@ public final class ASTMethodReference extends AbstractJavaExpr
         this.functionalMethod = methodType;
     }
 
-    void setCompileTimeDecl(JMethodSig methodType) {
-        this.compileTimeDecl = methodType;
+    void setCompileTimeDecl(OverloadSelectionResult ctdecl) {
+        this.compileTimeDecl = ctdecl;
     }
 
 }